+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Restructured mouse input patching so testing with UnityExplorer is easier
 - Made some changes to the arms in multiplayer
 - Reworked networking internals, should not affect gameplay
+- Reworked compatible version checking
 
 **Removals**:
 - Removed F8 hotswapping completely